Refactor schema tests into separate files...
One file contains tests that are used for testing the schema module
itself, and the other file contains tests used for testing actual ASDF
schemas.
